refactor: file names and dependencies
