+++ backend/src/auth/dto/index.ts
@@ -1,6 +1,6 @@
-import { TwoFactorActivatedDto } from './twofactorActivated.dto';
-import { TwoFactorCodeDto } from './twofactorCode.dto';
-import { TwoFactorValidDto } from './twofactorValid.dto';
+import { TwoFactorActivatedDto } from './two-factor.activated.dto';
+import { TwoFactorCodeDto } from './two-factor.code.dto';
+import { TwoFactorValidDto } from './two-factor.valid.dto';
 
 export const authDtos = [
   TwoFactorCodeDto,
